requirements.txt
crc-nd.utils==0.2.0
Django==1.6.7
django-tastypie==0.12.0
mock==1.0.1
path.py==5.3
psutil==2.1.2
requests==2.4.1
South==1.0
vecnet.simulation==0.1
